+++ src/main/java/io/github/astrapi69/gen/tree/enumeration/merge/MergeStrategy.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.gen.tree.merge.enumeration;
+package io.github.astrapi69.gen.tree.enumeration.merge;
 
 /**
  * The enum {@link MergeStrategy} will be used for merge tree nodes