+++ src/main/java/io/github/astrapi69/gen/tree/enumeration/traversal/TraversalType.java
@@ -1,0 +1,20 @@
+package io.github.astrapi69.gen.tree.enumeration.traversal;
+
+/**
+ * The enum {@link TraversalType} provides the different types of traversing a tree
+ */
+public enum TraversalType
+{
+	/**
+	 * This <code>INORDER</code> type is for the inorder traversal
+	 */
+	INORDER,
+	/**
+	 * This <code>PREORDER</code> type is for the preorder traversal
+	 */
+	PREORDER,
+	/**
+	 * This <code>POSTORDER</code> type is for the postorder traversal
+	 */
+	POSTORDER
+}